Convert dmraid to use dm-raid module for RAID4 and RAID5
The dmraid library used the dm-raid45 target for RAID4 and RAID5 sets.
This module however was never upstream and is suffering constant issues
when the Linux kernel changes.
Upstream introduced the dm-raid/raid module/target in 2.6.38. This can be
used to create device-mapper targets which use the MD RAID disciplines
(though not showing up in /proc/mdstat).

Most of the table constructor arguments can be mapped, only the member
devices may not have an offset.

From the old code it seems that RAID5 would always be build by set
elements which will not have an offset. So potentially only RAID4
could cause issues. That could be worked around if needed by creating
additional linear mappings.

NOTE: The event code is only compile tested. But we seem to have never
shipped the dmraid event tool which is needed to use this feature. And
we are about to remove support for ISW from dmraid (in favour of mdadm)
anyway.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>